Remove CameraMainView and other unneeded views
