+++ src/components/Views/CameraView/CameraView.test.js
@@ -4,6 +4,10 @@ import 'react-native';
 import renderer from 'react-test-renderer';
 import CameraView from './CameraView';
 
+jest.mock('../ObservationsView', () => () => null);
+jest.mock('../MapView', () => () => null);
+jest.mock('../../Base/SavedModal', () => () => null);
+
 describe('CameraView tests', () => {
   const isFocused = () => true;
   const addListener = () => true;
