+++ src/components/Views/CameraView/index.js
@@ -2,22 +2,38 @@
 import { connect } from 'react-redux';
 import { NavigationActions } from 'react-navigation';
 import type { Dispatch } from 'redux';
-import { observationUpdate } from '../../../ducks/observations';
+import {
+  observationCreate,
+  observationUpdate
+} from '../../../ducks/observations';
 import type { StoreState } from '../../../types/redux';
 import CameraView from './CameraView';
-import type { StateProps, DispatchProps } from './CameraView';
+import type { Props, StateProps, DispatchProps } from './CameraView';
+import { drawerClose, drawerOpen } from '../../../ducks/drawers';
 
-function mapStateToProps(state: StoreState): StateProps {
+function mapStateToProps(state: StoreState, ownProps: Props): StateProps {
   return {
-    selectedObservation: state.app.selectedObservation
+    observations: state.app.observations,
+    selectedObservation: state.app.selectedObservation,
+    showSavedModal: state.app.modals.saved,
+    fromEditor:
+      ownProps.navigation &&
+      ownProps.navigation.state &&
+      ownProps.navigation.state.params &&
+      ownProps.navigation.state.params.source === 'editor'
   };
 }
 
 function mapDispatchToProps(dispatch: Dispatch): DispatchProps {
   return {
+    createObservation: observation => dispatch(observationCreate(observation)),
     updateObservation: observation => dispatch(observationUpdate(observation)),
     goToObservationEditor: () =>
-      dispatch(NavigationActions.navigate({ routeName: 'ObservationEditor' }))
+      dispatch(NavigationActions.navigate({ routeName: 'ObservationEditor' })),
+    goToCategories: () =>
+      dispatch(NavigationActions.navigate({ routeName: 'Categories' })),
+    onDrawerClose: () => dispatch(drawerClose('observations')),
+    onDrawerOpen: () => dispatch(drawerOpen('observations'))
   };
 }
