+++ src/components/Views/MapView/Map/Map.test.js
@@ -2,17 +2,18 @@
 import React from 'react';
 import 'react-native';
 import renderer from 'react-test-renderer';
-import Position from './Position';
-import { createObservation } from '../../../mocks/observations';
+import Map from './Map';
+import { createObservation } from '../../../../mocks/observations';
 
-describe('Position tests', () => {
+describe('MapView tests', () => {
   test('snapshots', () => {
     const observation = createObservation();
     const isFocused = () => true;
     const addListener = () => true;
 
-    const tree = renderer.create(
-      <Position
+    let tree;
+    tree = shallow(
+      <Map
         navigation={{ isFocused, addListener }}
         selectedObservation={observation}
       />
