+++ src/components/Views/MapView/index.js
@@ -1,42 +1,16 @@
 // @flow
 import { connect } from 'react-redux';
-import { NavigationActions } from 'react-navigation';
-import type { Dispatch } from 'redux';
-import {
-  observationList,
-  observationCreate,
-  observationUpdate,
-  observationSelect
-} from '../../../ducks/observations';
 
-import { StoreState } from '../../../types/redux';
 import MapView from './MapView';
-import type { DispatchProps, StateProps } from './MapView';
+import { drawerClose, drawerOpen } from '../../../ducks/drawers';
 
-const mapStateToProps = (state: StoreState): StateProps => ({
-  observations: state.app.observations,
-  selectedObservation: state.app.selectedObservation,
-  gps: state.app.gps.data
+const mapStateToProps = state => ({
+  showSavedModal: state.app.modals.saved
 });
 
-function mapDispatchToProps(dispatch: Dispatch): DispatchProps {
-  return {
-    listObservations: () => dispatch(observationList('')),
-    createObservation: observation => dispatch(observationCreate(observation)),
-    goToCategories: () =>
-      dispatch(
-        NavigationActions.navigate({
-          routeName: 'Categories',
-          key: 'CategoriesView'
-        })
-      ),
-    updateObservation: observation => dispatch(observationUpdate(observation)),
-    selectObservation: o => dispatch(observationSelect(o)),
-    goToObservationDetail: () =>
-      dispatch(
-        NavigationActions.navigate({ routeName: 'ObservationDetailView' })
-      )
-  };
-}
+const mapDispatchToProps = dispatch => ({
+  onDrawerClose: () => dispatch(drawerClose('observations')),
+  onDrawerOpen: () => dispatch(drawerOpen('observations'))
+});
 
 export default connect(mapStateToProps, mapDispatchToProps)(MapView);
