--- src/components/Views/Position/index.js
@@ -1,20 +1,0 @@
-// @flow
-import { connect } from 'react-redux';
-import type { Dispatch } from 'redux';
-
-import { observationUpdate } from '../../../ducks/observations';
-import type { StoreState } from '../../../types/redux';
-import Position from './Position';
-import type { StateProps, DispatchProps } from './Position';
-
-function mapStateToProps(state: StoreState): StateProps {
-  return { selectedObservation: state.app.selectedObservation };
-}
-
-function mapDispatchToProps(dispatch: Dispatch): DispatchProps {
-  return {
-    updateObservation: observation => dispatch(observationUpdate(observation))
-  };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(Position);
