+++ src/ducks/index.js
@@ -2,7 +2,6 @@ import { combineReducers } from 'redux';
 import { createNavigationReducer } from 'react-navigation-redux-helpers';
 import { combineReducers as appCombineReducers } from '../lib/redux';
 import MainStackNavigation from '../components/MainNavigation/MainStackNavigation';
-import TabBarNavigation from '../components/Views/TabBarNavigation';
 
 import categories from './categories';
 import drawers from './drawers';
