+++ src/models/observations.js
@@ -14,6 +14,7 @@ export const applyObservationDefaults = (partial: Object): Observation => ({
   icon: null,
   categoryId: '0',
   fields: [],
+  createdFrom: 'map',
 
   ...partial
 });
