+++ src/types/observation.js
@@ -23,4 +23,5 @@ export interface Observation {
   icon: any;
   categoryId: string;
   fields: Field[];
+  createdFrom: string;
 }
